fix: 미니게임 수정
